Add required field for operation parameters #25
For operations bound to an entity or an entity collection, operation
imports.

Field Name	Type
required	boolean
Determines whether this parameter is mandatory. If the parameter is in
"path", this property is required and its value MUST be true. Otherwise,
the property MAY be included and its default value is false.
https://github.com/swagger-api/swagger-spec/blob/master/versions/2.0.md#parameterObject